extract abstract validator
